# frontend/shared/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AgentConfigCommandsAckRequest": {
        "additionalProperties": false,
        "properties": {
          "appliedRevision": {
            "format": "int64",
            "type": "integer"
          },
          "commandId": {
            "format": "int64",
            "type": "integer"
          },
          "failureReason": {
            "type": "string"
          },
          "mirrorContent": {
            "$ref": "#/components/schemas/JsonObject"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "appliedRevision",
          "commandId",
          "mirrorContent",
          "status"
        ],
        "type": "object"
      },
      "AgentConfigCommandsPullRequest": {
        "additionalProperties": false,
        "properties": {
          "localRevision": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "localRevision"
        ],
        "type": "object"
      },
      "AgentConfigCommandsPullResponse": {
        "additionalProperties": false,
        "properties": {
          "deviceId": {
            "format": "int64",
            "type": "integer"
          },
          "mirrorContent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JsonObject"
              },
              {
                "type": "null"
              }
            ]
          },
          "pendingCommands": {
            "items": {
              "$ref": "#/components/schemas/DeviceConfigCommandItem"
            },
            "type": "array"
          },
          "revision": {
            "format": "int64",
            "type": "integer"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "deviceId",
          "pendingCommands",
          "revision",
          "updatedAt"
        ],
        "type": "object"
      },
      "AgentConfigMirrorRequest": {
        "additionalProperties": false,
        "properties": {
          "localRevision": {
            "format": "int64",
            "type": "integer"
          },
          "mirrorContent": {
            "$ref": "#/components/schemas/JsonObject"
          },
          "summary": {
            "type": "string"
          }
        },
        "required": [
          "localRevision",
          "mirrorContent"
        ],
        "type": "object"
      },
      "AgentRegisterRequest": {
        "additionalProperties": false,
        "properties": {
          "appVersion": {
            "type": "string"
          },
          "bindCode": {
            "type": "string"
          },
          "deviceModel": {
            "type": "string"
          },
          "deviceName": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          }
        },
        "required": [
          "appVersion",
          "bindCode",
          "deviceModel",
          "deviceName",
          "platform"
        ],
        "type": "object"
      },
      "AgentRegisterResponse": {
        "additionalProperties": false,
        "properties": {
          "deviceId": {
            "format": "int64",
            "type": "integer"
          },
          "deviceToken": {
            "type": "string"
          },
          "userId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "deviceId",
          "deviceToken",
          "userId"
        ],
        "type": "object"
      },
...
      "BootstrapAdminRequest": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "password",
          "username"
        ],
        "type": "object"
      },
      "BootstrapAdminResponse": {
        "additionalProperties": false,
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "userId": {
            "format": "int64",
            "type": "integer"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "userId",
          "username"
        ],
        "type": "object"
      },
      "ChangePasswordRequest": {
        "additionalProperties": false,
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "type": "object"
      },
      "DesktopExchangeRequest": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "type": "object"
      },
      "DesktopLoginPageRequest": {
        "additionalProperties": false,
        "properties": {
          "client_name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "redirect_uri": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "client_name",
          "password",
          "redirect_uri",
          "state",
          "username"
        ],
        "type": "object"
      },
      "DesktopLogoutRequest": {
        "additionalProperties": false,
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "type": "object"
      },
      "DesktopRefreshRequest": {
        "additionalProperties": false,
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "type": "object"
      },
      "DesktopSessionResponse": {
        "additionalProperties": false,
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "authenticated": {
            "type": "boolean"
          },
          "expiresAt": {
            "type": "string"
          },
          "refreshExpiresAt": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "authenticated",
          "expiresAt",
          "refreshExpiresAt",
          "refreshToken",
          "username"
        ],
        "type": "object"
      },
...
      "DeviceConfigCommandItem": {
        "additionalProperties": false,
        "properties": {
          "actorId": {
            "format": "int64",
            "type": "integer"
          },
          "actorType": {
            "type": "string"
          },
          "appliedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "baseRevision": {
            "format": "int64",
            "type": "integer"
          },
          "createdAt": {
            "type": "string"
          },
          "failureReason": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "mutation": {
            "$ref": "#/components/schemas/JsonObject"
          },
          "status": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "targetRevision": {
            "format": "int64",
            "type": "integer"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "actorId",
          "actorType",
          "baseRevision",
          "createdAt",
          "id",
          "mutation",
          "status",
          "summary",
          "targetRevision",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "DeviceConfigStateResponse": {
        "additionalProperties": false,
        "properties": {
          "deviceId": {
            "format": "int64",
            "type": "integer"
          },
          "mirrorContent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JsonObject"
              },
              {
                "type": "null"
              }
            ]
          },
          "pendingCommands": {
            "items": {
              "$ref": "#/components/schemas/DeviceConfigCommandItem"
            },
            "type": "array"
          },
          "revision": {
            "format": "int64",
            "type": "integer"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "deviceId",
          "pendingCommands",
          "revision",
          "updatedAt"
        ],
        "type": "object"
      },
      "DeviceItem": {
        "additionalProperties": false,
        "properties": {
          "appVersion": {
            "type": "string"
          },
          "capabilities": {
            "$ref": "#/components/schemas/JsonObject"
          },
          "createdAt": {
            "type": "string"
          },
          "deviceModel": {
            "type": "string"
          },
          "deviceName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "lastSeenAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "localAddresses": {
            "$ref": "#/components/schemas/JsonObject"
          },
          "platform": {
            "type": "string"
          },
          "revokedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedAt": {
            "type": "string"
          },
          "userId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "appVersion",
          "capabilities",
          "createdAt",
          "deviceModel",
          "deviceName",
          "displayName",
          "enabled",
          "id",
          "localAddresses",
          "platform",
          "updatedAt",
          "userId"
        ],
        "type": "object"
      },
...
      "HeartbeatRequest": {
        "additionalProperties": false,
        "properties": {
          "appVersion": {
            "type": "string"
          },
          "capabilities": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object"
          },
          "localAddresses": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "appVersion",
          "capabilities",
          "localAddresses"
        ],
        "type": "object"
      },
      "JsonObject": {
        "additionalProperties": true,
        "type": "object"
      },
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "password",
          "username"
        ],
        "type": "object"
      },
      "LoginResponse": {
        "additionalProperties": false,
        "properties": {
          "authenticated": {
            "type": "boolean"
          },
          "csrfToken": {
            "type": "string"
          },
          "languageTag": {
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "authenticated",
          "csrfToken",
          "username"
        ],
        "type": "object"
      },
...
      "PatchDeviceRequest": {
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "enabled": {
            "type": [
              "boolean",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "RecordsResponse": {
        "additionalProperties": false,
        "properties": {
          "limit": {
            "format": "int64",
            "type": "integer"
          },
          "offset": {
            "format": "int64",
            "type": "integer"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/RelayRecord"
            },
            "type": "array"
          }
        },
        "required": [
          "limit",
          "offset",
          "records"
        ],
        "type": "object"
      },
      "RelayRecord": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "type": "string"
          },
          "callType": {
            "format": "int64",
            "type": "integer"
          },
          "deviceId": {
            "format": "int64",
            "type": "integer"
          },
          "eventId": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "metadata": {
            "$ref": "#/components/schemas/JsonObject"
          },
          "msgType": {
            "format": "int64",
            "type": "integer"
          },
          "occurredAt": {
            "type": "string"
          },
          "packageName": {
            "type": "string"
          },
          "recordType": {
            "type": "string"
          },
          "sender": {
            "type": "string"
          },
          "smsCode": {
            "type": "string"
          },
          "uploadedAt": {
            "type": "string"
          }
        },
        "required": [
          "body",
          "callType",
          "deviceId",
          "id",
          "metadata",
          "msgType",
          "occurredAt",
          "packageName",
          "recordType",
          "sender",
          "smsCode",
          "uploadedAt"
        ],
        "type": "object"
      },
...
      "RelayRecordsBatchRequest": {
        "additionalProperties": false,
        "properties": {
          "records": {
            "items": {
              "$ref": "#/components/schemas/RelayRecordWire"
            },
            "type": "array"
          }
        },
        "required": [
          "records"
        ],
        "type": "object"
      },
      "RelayRecordsBatchResponse": {
        "additionalProperties": false,
        "properties": {
          "inserted": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "inserted"
        ],
        "type": "object"
      },
      "SimpleOKResponse": {
        "additionalProperties": false,
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Contract for the Xinyi Relay backend, Android agent, Web console, and Desktop console.",
    "title": "Xinyi Relay Remote API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/agent/config/commands:ack": {
      "post": {
        "operationId": "ackAgentConfigCommands",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentConfigCommandsAckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceConfigCommandItem"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/agent/config/commands:pull": {
      "post": {
        "operationId": "pullAgentConfigCommands",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentConfigCommandsPullRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConfigCommandsPullResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/agent/config/mirror": {
      "post": {
        "operationId": "pushAgentConfigMirror",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentConfigMirrorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceConfigStateResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/agent/heartbeat": {
      "post": {
        "operationId": "sendAgentHeartbeat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleOKResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/agent/records:batch": {
      "post": {
        "operationId": "ingestAgentRecords",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelayRecordsBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelayRecordsBatchResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/agent/register": {
      "post": {
        "operationId": "registerAgent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRegisterResponse"
                }
              }
            },
            "description": "Created"
          }
        }
      }
    },
    "/api/v1/auth/desktop/exchange": {
      "post": {
        "operationId": "exchangeDesktopAuthCode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DesktopExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesktopSessionResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/auth/desktop/logout": {
      "post": {
        "operationId": "logoutDesktopSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DesktopLogoutRequest"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleOKResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/auth/desktop/refresh": {
      "post": {
        "operationId": "refreshDesktopSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DesktopRefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesktopSessionResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/auth/desktop/start": {
      "get": {
        "operationId": "getDesktopAuthStartPage",
        "parameters": [
          {
            "in": "query",
            "name": "redirect_uri",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "client_name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "OK"
          }
        }
      },
      "post": {
        "operationId": "approveDesktopAuthStart",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/DesktopLoginPageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "303": {
            "description": "Desktop auth callback redirect"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/auth/password": {
      "post": {
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleOKResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/bootstrap/admin": {
      "post": {
        "operationId": "bootstrapAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BootstrapAdminRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapAdminResponse"
                }
              }
            },
            "description": "Created"
          }
        }
      }
    },
...
    "/api/v1/devices/{id}": {
      "patch": {
        "operationId": "patchDevice",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceItem"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/records": {
      "get": {
        "operationId": "listRecords",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "device_id",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordsResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/records/{id}": {
      "get": {
        "operationId": "getRecord",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelayRecord"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "description": "Same-origin backend",
      "url": "/"
    }
  ]
}